+++ packages/wonder-blocks-icon-button/src/index.ts
@@ -1,1 +1,2 @@
 export {IconButton as default} from "./components/icon-button";
+export {ActivityIconButton} from "./components/activity-icon-button";